% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FGFR3_signature.R
\name{FGFR3_signature}
\alias{FGFR3_signature}
\title{Calculate the FGFR3 Signature Score}
\usage{
FGFR3_signature(expr)
}
\arguments{
\item{expr}{A matrix or data frame of gene expression values with genes as rows and samples as columns.}
}
\value{
A data frame with FGFR3 signature scores for each sample.
}
\description{
Calculate the FGFR3 Signature Score
}
